requirements.txt
langgraph==0.1.7
langchain>=0.2.11,<0.3.0
langchain-google-genai>=1.0.1,<2.0.0
google-generativeai>=0.7.0,<0.8.0
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
h11==0.14.0
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
celery==5.3.4
websockets==12.0
watchfiles==0.21.0